requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
cssselect2==0.6.0
daphne==3.0.2
dj-database-url==1.0.0
Django==4.1.1
django-ckeditor==6.5.1
django-embed-video==1.4.5
django-js-asset==2.0.0
django-xhtml2pdf==0.0.4
future==0.18.2
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
lxml==4.9.1
oscrypto==1.3.0
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.0.0
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
svglib==1.4.1
tinycss2==1.1.1
tqdm==4.64.1
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.8
zope.interface==5.4.0
